requirements.txt
python-dotenv
redditharbor
